product_scraper/requirements.txt
Pillow>=11.0.0
playwright>=1.49.0
httpx>=0.27.0
